moved Hypertrie out of interface folder
