+++ include/Dice/hypertrie/hypertrie.hpp
@@ -1,0 +1,8 @@
+#ifndef HYPERTRIE_HYPERTRIE_HPP
+#define HYPERTRIE_HYPERTRIE_HPP
+
+#include "Dice/hypertrie/internal/Hypertrie.hpp"
+#include "Dice/hypertrie/internal/HashJoin.hpp"
+
+
+#endif//HYPERTRIE_HYPERTRIE_HPP
