+++ include/Dice/hypertrie/internal/ConfigHypertrieDepthLimit.hpp
@@ -3,6 +3,7 @@
 #include <cctype>
 
 namespace hypertrie::internal {
+	// TODO: add option to override this at startup
 static constexpr const std::size_t hypertrie_depth_limit = 6;
 }
 #endif//HYPERTRIE_CONFIGHYPERTRIEDEPTHLIMIT_HPP
