+++ include/Dice/hypertrie/internal/Hypertrie.hpp
@@ -3,10 +3,10 @@
 
 #include "Dice/hypertrie/internal/container/AllContainer.hpp"
 
-#include "Dice/hypertrie/internal/interface/HashDiagonal.hpp"
-#include "Dice/hypertrie/internal/interface/HypertrieContext.hpp"
-#include "Dice/hypertrie/internal/interface/Hypertrie_traits.hpp"
-#include "Dice/hypertrie/internal/interface/Iterator.hpp"
+#include "Dice/hypertrie/internal/HashDiagonal.hpp"
+#include "Dice/hypertrie/internal/HypertrieContext.hpp"
+#include "Dice/hypertrie/internal/Hypertrie_traits.hpp"
+#include "Dice/hypertrie/internal/Iterator.hpp"
 
 #include "Dice/hypertrie/internal/util/CONSTANTS.hpp"
 #include <optional>
